telegram_bot/requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
matplotlib==3.8.2
